Ready For review pending testing on physical hardware

+++ firmware/Core/Src/antenna_deploy_drivers/ant_commands.c
@@ -181,11 +181,12 @@ uint8_t ANT_CMD_measure_temp(uint16_t *result) {
     }
     return comms_err;
 }
-/// @brief  Converts temperature measurement to degrees celsius
+/// @brief  Converts temperature measurement to centi-degrees celsius
 /// @param measurement measurement result from ANT_CMD_measure_temp()
-/// @return returns the temperature in centi degrees celsius
-int32_t ANT_CMD_convert_temp_measurement_to_centi_degree_celsius(uint16_t measurement) {
-    return (int32_t)(((float) measurement * -0.29648 + 193.375) * 100);
+/// @return returns the temperature in centi-degrees celsius
+/// @note The formula used is based on the formula found in the data sheet for the LM94022. Derivation can be found on github issue#146.
+int16_t ANT_CMD_convert_temp_measurement_to_centi_degree_celsius(uint16_t measurement) {
+    return (int16_t)(((float) measurement * -0.29648 + 193.375) * 100);
 }
